refactor enums package 2.0
