+++ src/main/java/edu/lucasrech/catalog_authentication/model/product/Product.java
@@ -2,6 +2,7 @@ package edu.lucasrech.catalog_authentication.model.product;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import edu.lucasrech.catalog_authentication.model.product.enums.Category;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
