+++ src/main/java/edu/lucasrech/catalog_authentication/model/product/ProductDTO.java
@@ -1,5 +1,7 @@
 package edu.lucasrech.catalog_authentication.model.product;
 
 
+import edu.lucasrech.catalog_authentication.model.product.enums.Category;
+
 public record ProductDTO(String name, String description, double price, String image_url, Category category) {
 }
