+++ src/main/java/edu/lucasrech/catalog_authentication/model/user/User.java
@@ -1,5 +1,6 @@
 package edu.lucasrech.catalog_authentication.model.user;
 
+import edu.lucasrech.catalog_authentication.model.user.enums.Role;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
 import lombok.*;
