+++ src/main/java/edu/lucasrech/catalog_authentication/repository/ProductRepository.java
@@ -1,7 +1,7 @@
 package edu.lucasrech.catalog_authentication.repository;
 
 import edu.lucasrech.catalog_authentication.model.product.Product;
-import edu.lucasrech.catalog_authentication.model.product.Category;
+import edu.lucasrech.catalog_authentication.model.product.enums.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
